+++ network/packet18enablecard.h
@@ -12,7 +12,7 @@ public:
 
 public slots:
     void bytesToRead(QIODevice*, ClientSocket*);
-    void bytesToWrite(ClientSocket*c, quint8 card);
+    void bytesToWrite(ClientSocket*c, quint16 card);
 };
 
 #endif // PACKET18ENABLECARD_H
